added removeGroup(), removed LOSE.fmode...
LOSE.fmode was only used by a single function so it was removed from the LOSE class and moved to that function as a keyword argument.
LOSE.get_shape() was renamed to LOSE.getShape().
LOSE.make_generator() was renamed to LOSE.makeGenerator() and had some internal improvments.
LOSE.save() now only works in append mode.

check the README.md for more info on the changes.